+++ crates/net/network/src/transactions/config.rs
@@ -251,80 +251,3 @@ where
 /// Type alias for `TypedRelaxedFilter`. This filter accepts known Ethereum transaction types and
 /// ignores unknown ones without penalizing the peer.
 pub type RelaxedEthAnnouncementFilter = TypedRelaxedFilter<TxType>;
-
-/// A bundle of policies that control the behavior of network components like
-/// the [`TransactionsManager`](super::TransactionsManager).
-///
-/// This trait allows for different collections of policies to be used interchangeably.
-pub trait Policies: Send + Sync + Debug + 'static {
-    /// The type of the policy used for transaction propagation.
-    type Propagation: TransactionPropagationPolicy;
-    /// The type of the policy used for filtering transaction announcements.
-    type Announcement: AnnouncementFilteringPolicy;
-
-    /// Returns a reference to the transaction propagation policy.
-    fn propagation_policy(&self) -> &Self::Propagation;
-
-    /// Returns a mutable reference to the transaction propagation policy.
-    fn propagation_policy_mut(&mut self) -> &mut Self::Propagation;
-
-    /// Returns a reference to the announcement filtering policy.
-    fn announcement_filter(&self) -> &Self::Announcement;
-}
-
-/// A container that bundles specific implementations of transaction-related policies,
-///
-/// This struct implements the [`Policies`] trait, providing a complete set of
-/// policies required by components like the [`TransactionsManager`](super::TransactionsManager).
-/// It holds a specific [`TransactionPropagationPolicy`] and an
-/// [`AnnouncementFilteringPolicy`].
-#[derive(Debug, Clone)]
-pub struct NetworkPolicies<P, A>
-where
-    P: TransactionPropagationPolicy,
-    A: AnnouncementFilteringPolicy,
-{
-    propagation: P,
-    announcement: A,
-}
-
-impl<P, A> NetworkPolicies<P, A>
-where
-    P: TransactionPropagationPolicy,
-    A: AnnouncementFilteringPolicy,
-{
-    /// Creates a new bundle of network policies.
-    pub const fn new(propagation: P, announcement: A) -> Self {
-        Self { propagation, announcement }
-    }
-}
-
-impl<P, A> Policies for NetworkPolicies<P, A>
-where
-    P: TransactionPropagationPolicy + Debug,
-    A: AnnouncementFilteringPolicy + Debug,
-{
-    type Propagation = P;
-    type Announcement = A;
-
-    fn propagation_policy(&self) -> &Self::Propagation {
-        &self.propagation
-    }
-
-    fn propagation_policy_mut(&mut self) -> &mut Self::Propagation {
-        &mut self.propagation
-    }
-
-    fn announcement_filter(&self) -> &Self::Announcement {
-        &self.announcement
-    }
-}
-
-impl Default for NetworkPolicies<TransactionPropagationKind, StrictEthAnnouncementFilter> {
-    fn default() -> Self {
-        Self {
-            propagation: TransactionPropagationKind::default(),
-            announcement: StrictEthAnnouncementFilter::default(),
-        }
-    }
-}
